Backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
pymongo==4.10.1
python-dotenv==1.0.1
